One Last Edit
Deleted PrioritizeBishopMoves and PrioritizeKnightMoves and any tests associated with them. PrioritizePawnMoves has been left untouched in case you need to refer back to it for any reason.

Unfortunately, I have a series of commitments for the next 6 weeks that will prevent me from continuing to work on the project. If you still need help come August, I would be happy to return to chessmate. Until then, best of luck with your program! :-)